% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medusaRClient.read.pmlame.R
\name{medusaRClient.read.pmlame}
\alias{medusaRClient.read.pmlame}
\title{Get pmlame json from medusa and return a dataframe of pmlame}
\usage{
medusaRClient.read.pmlame(pmlame_or_stone, opts = NULL, verbose = TRUE)
}
\arguments{
\item{pmlame_or_stone}{A dataframe that already converted or
stone-ID.}

\item{opts}{List of further options for plot.
options are:
 - 'Recursivep' is an option whether to get whole family's information.}

\item{verbose}{Output debug info (default: TRUE).}
}
\value{
A dataframe with unit organized.
  \url{https://github.com/misasa/MedusaRClient}
}
\description{
Get pmlame json from medusa and return a dataframe of pmlame.
}
\details{
This function converts a json data which is from medusa ([medusa_host]/records/[global_id]/pmlame.json) to a dataframe and read it.
}
\examples{
global_id <- "20081202172326.hkitagawa"
pmlame <- medusaRClient.read.pmlame(global_id)
}
