% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmlame.R
\docType{class}
\name{Pmlame}
\alias{Pmlame}
\title{Return an R object that will return dataframe or dataframe}
\format{\code{\link{R6Class}} object.}
\usage{
Pmlame
}
\value{
An R object of \code{\link{R6Class}} with methods that
  communicate with Medusa or dataframe.
}
\description{
Return an R object that will return dataframe or
  dataframe.  This class provides a coherent wrapper
  object-relational mapping for Medusa's RESTful Web service.
}
\section{Methods}{

\describe{
  \item{Documentation}{Return an R object that will return dataframe or dataframe.
    Visit \url{https://github.com/misasa/MedusaRClient/} for detail.}
  \item{\code{new()}}{Return a new R object that corresponds to a port (or a connection) for dataframes.}
  \item{\code{read(global_id, options = NULL)}}{Return a dataframe of
    a record with \code{global_id} specified.  Return a recursive dataframe when
    \code{Recursivep} is TRUE.}
}
}

\examples{
global_id <- "20081202172326.hkitagawa"
conn <- Connection$new(list(uri="devel.misasa.okayama-u.ac.jp/Chelyabinsk/", user="admin", password="admin"))
obj <- Pmlame$new(conn)
pmlame <- obj$read(global_id, list(Recursivep=TRUE))
}
\keyword{datasets}
