% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\docType{class}
\name{Resource}
\alias{Resource}
\title{Return an R object that will return Medusa records or Medusa records}
\format{\code{\link{R6Class}} object.}
\usage{
Resource
}
\value{
An R object of \code{\link{R6Class}} with methods that
  communicate with Medusa or R objects that correspond to Medusa records.
}
\description{
Return an R object that will return Medusa records or
  Medusa records.  This class provides a coherent wrapper
  object-relational mapping for Medusa's RESTful web service.
}
\section{Fields}{

\describe{
\item{\code{name}}{Stores resource name of your Medusa server}

\item{\code{conn}}{Stores connection object}
}}

\section{Methods}{

\describe{
  \item{Documentation}{Return an R object that will return Medusa records or Medusa records.
  Visit \url{https://github.com/misasa/MedusaRClient/} for detail.}
  \item{\code{new(NAME)}}{Return a new R object that corresponds to records having resource
  name \code{NAME} in Medusa.  As of March 2018, "specimens", "boxes", "places", "surfaces",
  "spots", "analyses", "devices", "techniques", "chemistries", "bibs", and "attachment_files" are defined in Medusa.}
  \item{\code{find(id, options = NULL)}}{Return a Medusa record with local \code{ID} specified.}
  \item{\code{find_by_ids(ids, options = NULL)}}{Return Medusa records with local \code{ID} specified.}
  \item{\code{find_by_global_id(id, options = NULL)}}{Return a Medusa record with glocal \code{ID} specified.}
  \item{\code{find_all(options = NULL)}}{Return all Medusa records having the resource name.}
}
}

\examples{
Specimen <- Resource$new("specimens")
Specimen$find_all()
Box <- Resource$new("boxes")
Place <- Resource$new("places")
surface <- Resource$new("surfaces")
Spot <- Resource$new("spots")
Analysis <- Resource$new("analyses")
Device <- Resource$new("devices")
Technique <- Resource$new("techniques")
Chemistry <- Resource$new("chemistries")
Bib <- Resource$new("bibs")
AttachmentFile <- Resource$new("attachment_files")
}
\seealso{
\code{\link{connection}}
}
\keyword{datasets}
