% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{class}
\name{Connection}
\alias{Connection}
\title{Return an R object that will return association lists or association lists}
\format{\code{\link{R6Class}} object.}
\usage{
Connection
}
\value{
An R object of \code{\link{R6Class}} with methods that
  communicate with Medusa or association lists.
}
\description{
Return an R object that will return association lists
  or association lists.  This class provides object with methods
  for communication with RESTful services.  This is low-level
  function and it is not recommended for an user to call this
  function.
}
\section{Methods}{

\describe{
  \item{Documentation}{Return an R object that will return association lists
  or association lists.  Visit \url{https://github.com/misasa/MedusaRClient/} for detail.}
  \item{\code{new()}}{Return a new R object that corresponds to a port (or a connection) for association lists.}
  \item{\code{get(resource, id, params = NULL)}}{Get record with \code{resource} as resource name, \code{id} as id, and \code{params} as list of query parameters}
  \item{\code{gets(resource, ids, params = NULL)}}{Get record with \code{resource} as resource name, \code{ids} as list of ids, and \code{params} as list of query parameters}
  \item{\code{gets_sub(owner, id, resource, params = NULL)}}{Get record with \code{owner} as parent resource name, \code{id} as id of parent record, \code{resource} as resource name, and \code{params} as list of query parameters}
}
}

\examples{
conn <- Connection$new(list(uri="devel.misasa.okayama-u.ac.jp/Chelyabinsk/", user="admin", password="admin"))
record <- conn$get("specimens",23952)
}
\seealso{
\code{\link{pmlame}} and \code{\link{resource}}
}
\keyword{datasets}
