% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmlame.R
\docType{class}
\name{Pmlame}
\alias{Pmlame}
\title{Class providing object with methods for communication with Medusa server}
\format{\code{\link{R6Class}} object.
#' @section Methods:
\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/misasa/MedusaRClient/}
  \item{\code{new()}}{This method is used to create object of the class}
  \item{\code{read(global_id, options = NULL)}{This method get a record with \code{global_id} as global-ID of the record and \code{opts} as Recursivep}}
}}
\usage{
Pmlame
}
\value{
Object of \code{\link{R6Class}} with methods for communication with Medusa server
}
\description{
Class providing object with methods for communication with Medusa server
}
\examples{
global_id <- "20081202172326.hkitagawa"
con <- Connection$new(list(uri="devel.misasa.okayama-u.ac.jp/Chelyabinsk/", user="admin", password="admin"))
obj <- Pmlame$new(con)
pmlame <- obj$read(global_id, list(Recursivep=TRUE))
}
\keyword{datasets}
