% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource.R
\docType{class}
\name{Resource}
\alias{Resource}
\title{Class providing object with methods for communication with Medusa server}
\format{\code{\link{R6Class}} object.}
\usage{
Resource
}
\value{
Object of \code{\link{R6Class}} with methods for communication with Medusa server
}
\description{
Class providing object with methods for communication with Medusa server
}
\section{Fields}{

\describe{
\item{\code{name}}{Stores resource name of your Medusa server}

\item{\code{conn}}{Stores connection object
#' @section Methods:
\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/misasa/MedusaRClient/}
  \item{\code{new(name)}}{This method is used to create object of the class with \code{name} as resource name of the server object}
  \item{\code{find(id, options = NULL)}{This method finds a record}}
  \item{\code{find_by_ids(ids, options = NULL)}{This methods finds records}}
  \item{\code{find_by_global_id(id, options = NULL)}{This method finds a record}}
  \item{\code{find_all(options = NULL)}{This method finds records}}
}}
}}

\examples{
Specimen <- Resource$new("specimens")
Specimen$find_all()
Box <- Resource$new("boxes")
Place <- Resource$new("places")
surface <- Resource$new("surfaces")
Spot <- Resource$new("spots")
Analysis <- Resource$new("analyses")
Device <- Resource$new("devices")
Technique <- Resource$new("techniques")
Chemistry <- Resource$new("chemistries")
Bib <- Resource$new("bibs")
AttachmentFile <- Resource$new("attachment_files")
}
\keyword{datasets}
